requirements.txt
flask==3.0.0
flask-cors==4.0.0
opencv-python-headless==4.8.1.78
numpy==1.26.0
openai==1.3.0
Pillow==10.1.0
shapely==2.0.2
mapbox_earcut==1.0.1
numpy-stl==3.1.0
scipy==1.11.3
